Bug 1198458: Webrtc updated to branch 43; pull made 2015-09-29 09:00AM PDT rs=jesup https://chromium.googlesource.com/external/webrtc git checkout -b my_branch refs/remotes/branch-heads/43 commit 6979b024d7cebfdcd1e8f66da59ea157bbc9e47e (I believe)

+++ media/webrtc/trunk/webrtc/modules/audio_coding/neteq/dtmf_buffer.h
@@ -69,7 +69,7 @@ class DtmfBuffer {
   // |rtp_timestamp| is simply copied into the struct.
   static int ParseEvent(uint32_t rtp_timestamp,
                         const uint8_t* payload,
-                        int payload_length_bytes,
+                        size_t payload_length_bytes,
                         DtmfEvent* event);
 
   // Inserts |event| into the buffer. The method looks for a matching event and
