+++ media/webrtc/trunk/webrtc/modules/video_coding/main/test/video_source.h
@@ -69,7 +69,7 @@ public:
     // Returns the filename with the path (including the leading slash) removed.
     std::string GetName() const;
 
-    int32_t GetFrameLength() const;
+    size_t GetFrameLength() const;
 
 private:
     std::string         _fileName;
